Add lambda with zip package support (WIP)
NOTE: this commit is result of a squash with multiple changes over
several weeks and attempts to get lambda dependencies to work correctly.
It was found that during analysis of issue #18, that it is not possible
to use .ZIP dependencies with lxml, bs4 library (see #18 for details and
links).

THIS WORK IS EXPERIMENTAL AND DOES NOT FUNCTION CORRECTLY!

This commit is created for educational purposes and eventually could be
used apply .ZIP dependencies approach instead of docker to deploy Python
dependencies.

Changes have been performed in following areas:
1. Change pylint config
2. README.md to update with steps/instructions to do it
3. Renaming requirements*.txt files, to split build and dev dependencies
   and fixes to GitHub actions
4. Makefile changes (in root and examples/zip directory)
5. Uses AWS_ACCOUNT environment variable to identify correctly AWS
   account
6. Depends on usage of `pip-compile` to fix issues wiht conflicting
   python dependencies