Remove docker implementation with alt. base image
Source: https://docs.aws.amazon.com/lambda/latest/dg/python-image.html#python-image-clients

As per above source, there was an attempt to use python-slim docker
image as base, to decrease size of it. Initially it was plausible idea,
however it proven to be more difficult to maintain, and operationally
not efficient..

It was decided, that it is preferred to keep simpler dockerfiles and
invokation, in expense of larger size of the images.